+++ Inventario/Application/ViewModels/CommissionViewModel.cs
@@ -15,9 +15,9 @@ namespace Application.ViewModels
         public ClientViewModel Client { get; set; }
         public string ClientName { get; set; }
         public string Product { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public ePaymentTypes PaymentType { get; set; }
         public DateTime Date { get; set; }
-        public double Value { get; set; }
+        public decimal Value { get; set; }
     }
 }
